Refactor API tests
